Allow destroying (and recreation) of GladeApp object
That is what happens when a glade file is loaded in anjuta, then closes and opened again.
Currently, we hold an extra reference on GladeApp to avoid destroying it but that's not clean
of course and leaves a big mem leak.
We cannot free the catalogs though as they register GTypes that we cannot unregister. Thus,
we have to keep them in memory once they were loaded.