Implement reading/writing of <binding> tags
This is the first in a (hopefully) long line of commits to implement
support for GObject property binding in Glade. It extends Glade's
GtkBuilder parsing/writing code to recognize property binding definitions
(which are codified as <binding> tags), create objects of the new
GladeBinding class to represent them, and write them back to the GtkBuilder
file when saving.

What this means is that if you now open a file with <binding> tags and
save them again later, the binding information is not lost.