Big refactor: Move all GladeBinding functionality directly into GladeProperty
As discussed with Tristan van Berkom. This fixes races that formerly existed
between creating a binding and attaching it to a GladeProperty (in this timeframe,
two GladeBindings were active at once) and - surprisingly - makes the code quite a
bit cleaner. This should be a much better base for implementing the next features
on my TODO list than the former one was.

As acherry on top, the new structure allowed me to fix a bug on the way: when
undoing a property binding action, the binding target property was editable
again, but the binding itself didn't actually disappear (meaning that editing
the source property still overrode the value of the target). Now, everything
works as intended. :)