Add support for setting transformation functions from "Bind to source..."
There is now a new checkbutton in the "Bind to source..." dialog to set
a transformation for a binding. Because we don't have the transformation
function and the source and target properties might have different types,
there is no in-workspace preview of the binding; instead, a property with
a transformed binding is shown as having its default value.