Deal with binding sourc properties becoming disabled or insensitive
Now, when the binding source of a property is disabled, the property
is reset to its default value and unbound. Enabling the source again
rebinds the property. Such "dead" property bindings are not saved
and thus only persist as long as Glade is opened. I hope this behavior
is in line in what users would expect; if not, it's easy to change.