Move all binding source/target synchronization logic into the command system
As to Tristan's request. This makes all the signal handling hackery I did
in glade-property.c obsolete, which greatly simplifies the code! The only
drawback is that changes of a binding source property's "enabled"/"sensitive"
state are not recognized anymore as these are not registered as commands.
How this will be fixed will still have to be discussed wih Pablo and Tristan
(likely after GSoC).