Rework the last commit to only introduce GladeWidgetAdaptor->evaluate_property_sensitivity()
As discussed with Tristan van Berkom on the mailing list. Instead of
recording property sensitivity changes directly in the command system,
a "property invalidated" signal will be added to the "gbinding" branch
to make sure that property bindings with invalidated source properties
are properly removed.

What remains in this branch, though, is a new virtual function
GladeWidgetAdaptor::evaluate_property_sensitivity() which centralizes
all the property sensitivity management which is currently scattered
around the plugins/gtk+/ codebase. As the invalidated-source-property
issue will now be solved directly in the "gbinding" branch, this is now
a purely cosmetic work, but nevertheless I will continue it (and probably
rebase the gbinding branch to it).

As in the last commit, of the code in plugins/gtk+/ only the GtkEntry
adaptor has been ported to use evaluate_property_sensitivity(), but the
others will eventually follow.