Add "invalidate" signal to GladeProperty
This is supposed to be signalled whenever a property gets invalid,
e.g. when switching a GtkLabel to use a pixbuf instead of a stock
icon. Binding target properties use this to remove any property
bindings that become invalid this way.

The signal is emitted when glade_property_invalidate() is called.
Currently, the GladeEntryEditor does this, so binding a property
to an entry's "Text" and then changing the entry to use a "Buffer"
correctly removes the binding. More callers will follow.