GtkButton adaptor: Fix the response id property load and save
Make the response id optional in the UI. If the user never sets the
response ID, then it will not be serialized in <action-widgets> of
a dialog or info bar.

This is needed because some buttons added to the action are not
response widgets (for instance, the credits and licence pages of
GtkAboutDialog) so it's important to not serialize them as such
unless they are explicitly marked as response widgets.