Added Glade's first unit test (about time !)
This patch clears the way for unit testing, some environment variables
are added to ensure unit tests run on code that is not installed but
in tree.

The added test iterates over all widgets in the catalog, uses the
adaptor to create a GladeWidget and asserts that the GladeWidget and
it's internal object is finalized after unreferencing the GladeWidget.