GladeEditorProperty: Manage construct-only property changes.
This is tricky, we block the GladeProperty change signals
while committing an new property value. However when the
property is construct-only, the GladeWidget temporarily
exits the GladeProject while it reconstructs the instance.

With this patch we avoid unblocking the property change
signal after a commit, if the signal was disconnected
and reconnected as a consequence of applying the property value.