Implemented autosave and backups
Whenever the user saves, the last version of the file is backed up.

Whenever the project is modified and idle for a configurable amount of
seconds, the project is automatically saved.

At load time, if the automatically saved file is newer than the proposed
Glade file, Glade will popup a dialog and let the user choose to load
the autosaved version instead