Patch up bug 589337 - Fix by Michal 'hramrach' Suchanek <hramrach@gmail.com>
This is a bandaid which fixes anoyance in dialogs which appear when creating
variable widgets like HBox.

The spin editor is used in multiple places but only htese dialogs have a
default button so it is safe to set the spinbutton to activate default
unconditionally. Does not seem to break anything.