Fixed selection bug exposed by Bug #652655 fix.
We made Project selection restoration more robust by saving a list of selected
objects before rebuilding one of them and clearing it after the rebuild and
just before restoring the selection.

We still have a minor but obscure bug where the property editor of the
construct only property does not work after rebuilding the object unless
the move is moved out of the widget.

This really shows why we need to split model data from runtime objects!