Fixed bug 584381 "Saving as an existing file name except the extension overwrites without confirmation."
The problem was that if the selected filename did not had an extension we appended append a .glade extension
but did not check if the new filename existed or not.
Now we fixed this problem by selecting the file if it exist and running the dialog again, giving it a chance
to promt the user about overwriting.