pango attribute editor: don't write on cancel for color/font
Properties such as 'Font Description' or 'Foreground Color' can be set
via a dialog. The current implementation also writes those values if the
user cancels the dialog. This is especially weird when using the font
dialog which can leave you with values such as "<Enter Value 10".
This patch fixes that problem by only writing the valuse if the user
clicked select.