Install widget icons into the hicolor theme.
Documentation seems to suggest that applications should
be installing their icons into the theme, we also no longer
call gtk-update-icon-cache at install time (we used to).

I made this particular change in order to cause the widget
icons to show up properly in bundles, if this commit is wrong
then we need to figure out how it actually works that icons
show up in the palette (so we can reproduce that in the bundle).