Added GLADE_BUNDLED environment variable.
This environment variable is used to detect a bundled environment
for which Glade will run on various 64bit linux systems.

Currently we just avoid loading any catalogs except for the gtk+.xml
catalog in the case of a bundled environment (but there may be
some extra special cases to add, in order to ensure we don't integrate
with the unknown running system).