Bug 720585 - Fix order of customized property output.
Recently we split up all the adaptor code into separate files,
this resulted in the above bug since the GtkWidget adaptor writes
out some custom attributes in it's ->write_widget() function.

Moved these Atk/StyleClasses/Accelerators to ->write_widget_after()
so that they will be written after a widget's children.

This fixes the GtkToolBar adaptor so that when it writes it's
icon size after chaining up to the parent adaptor, it does so still
inside the <properties> section of a widget.