gladeui: fix a pair of printf security problems
In two cases, gladeui was doing the equivalent of

  printf (var);

instead of the correct (and safe):

  printf ("%s", var);

This was caught by clang's pedantic treatment of non-literal format
strings forcing me to add G_GNUC_PRINTF attributes to functions
operating on format strings (which in turn caught the problem).  Those
changes are also included here.

https://bugzilla.gnome.org/show_bug.cgi?id=720883