glade-editor-property: use literal format strings
Refactor glade_eprop_object_dialog_title() to make use of literal format
strings instead of storing the format string into a variable.

It is my opinion that this change also improves readability of the
function in question by reducing the complexity and redundancy of the
checks.

This fixes the build which was broken on clang due to the use of
-Werror=format-nonliteral.

https://bugzilla.gnome.org/show_bug.cgi?id=720884