GladeWidgetAdaptor: Fixed glade_widget_adaptor_object_get_children() for internal widgets.
Internal widgets are defined in Glade in the adaptor class that introduced them.
So in order to get internal widgets that are descendant of another internal widget
we need to walk up the adaptor hierarchy to find where internal widgets are defined.

I included this patch in the stable version because without it glade can not access
GtkDialog internal action_area if compiled with gtk master (version 3.13)