Bug 732575 - Changed the type hint on the "Edit Separately" window to 'utility'
This window has traditionally had a 'dialog' type so that window
managers don't treat it as a primary application window (don't
add minimize/maximize, don't list it in the alt-tab list, etc).

In light of bug 732575, where it seems some window managers have
stolen the 'close' button away from windows with the 'dialog' hint,
let's change it for a 'utility' hint and hope this is the more
appropriate hint.