osx: Use current gtk-mac-integration API
The #ifdef'd block did not compile at all because it used the old API
from gtk-mac-integration, and because it used GtkBuilder when the
surrounding code had been ported to use widget templates.

This fixes up the use of the API and adds a few widget variables to
GladeWindowPrivate which have to be bound to the template.

https://bugzilla.gnome.org/show_bug.cgi?id=738339