GtkHeaderBar: Fixed property sensitivity issues
The custom editor was calling glade_widget_property_set_sensitive()
directly which does not work with undo/redo.

Cleaned up the editor and adaptor so that the editor issues commands
and the adaptor->set_property() adjusts property sensitivity when
'show-close-button' and 'use-custom-title' properties change.

Also some minor changes to the gtk+.xml.in, no need for the
decoration layout property to be optional really, since the property
is a string and will not be saved if it's not set.