Don't clear selection too eagerly
When the current submenu in a GtkPopoverMenu is selected, and we
are editing its submenu child property, we get change notification
for the popovers visible-submenu property. Don't clear the selection
in this case because the visible submenu did not actually change
to a different widget, it just changed its name.