glade-gtk-notebook.c: Fixed warning about imbalanced children
Now that there is a new special child type, only remove "tab"
widgets when extracting the children. This ugly loop is there
to ensure that the right tab is matched with it's page when
changing the notebook children during undo/redo, and also to
handle the case where positions of notebook tabs are not
specified in the datamodel (yet) because the notebook is
being loaded from a file and the appearance of tabs in
the loaded file should be used.

In any case, this small commit fixes the undo/redo console
warnings when undoing/redoing the addition/removal of children
from the new start action / end action positions.