Reverting change from notebook-actions branch merge.
It seems Matthias was testing headerbars in dialogs as was
suggested in the WONTFIX'ed bug 757562. Reverting the enabling
of headerbars in dialogs.

Also updated the comment in gtk+.xml.in around the headerbar
disabling, originally it was because it was unsupported (yet),
now it is because dialogs own their headerbars.