Ugly hack to restore paned positions
When we maximize the window, usually an idle is enough until it's
safe to start manipulating the stored paned window positions.

Now, 200ms seems to be enough.

This fixes restoring of paned window positions, if you exit Glade
cleanly by quitting Glade, it should restore the app in the same
state it was (undocked window positions and paned positions)