Fix formatting / layout of project verification details
Two of the common deprecation messages are missing a newline,
causing the message display to be 1000's of pixels wide
The scrolled window displaying the messages is also not set
to expand as the window grows, leaving only three lines of
text visible by default. These bugs make the verification
message display almost unusable.

Signed-off-by: Daniel P. Berrange <berrange@redhat.com>